Move Component Type from core objects to opendaq
