+++ core/opendaq/component/include/opendaq/component_type.h
