+++ core/opendaq/device/include/opendaq/device_type_impl.h
@@ -15,7 +15,7 @@
  */
 
 #pragma once
-#include <coreobjects/component_type_impl.h>
+#include <opendaq//component_type_impl.h>
 #include <opendaq/component_type_builder_ptr.h>
 #include <opendaq/device_type.h>
