+++ core/opendaq/functionblock/include/opendaq/function_block_type.h
@@ -15,12 +15,12 @@
  */
 
 #pragma once
-#include <coreobjects/component_type.h>
+#include <opendaq/component_type.h>
 
 BEGIN_NAMESPACE_OPENDAQ
 
 /*#
- * [interfaceSmartPtr(IComponentType, GenericComponentTypePtr, "<coreobjects/component_type_ptr.h>")]
+ * [interfaceSmartPtr(IComponentType, GenericComponentTypePtr, "<opendaq/component_type_ptr.h>")]
  * [interfaceLibrary(IComponentType, "coreobjects")]
  */
