+++ core/opendaq/server/include/opendaq/server_type_impl.h
@@ -16,7 +16,7 @@
 
 #pragma once
 #include <opendaq/server_type.h>
-#include <coreobjects/component_type_impl.h>
+#include <opendaq/component_type_impl.h>
 #include <opendaq/component_type_builder_ptr.h>
 
 BEGIN_NAMESPACE_OPENDAQ
